+++ plugins/custom-properties/custom-properties.js
@@ -6,7 +6,7 @@
 	
 	'use strict'
 
-	var properties = null
+	var store = null
 
 	function replace (match, group) {
 		return properties[group] || value
@@ -14,6 +14,16 @@
 
 	function customProperties (context, content, selectors, parents, line, column, length) {
 		switch (context) {
+			case -1: {
+				// create store
+				store = {}
+				break
+			}
+			case -2: {
+				// destroy store
+				store = null
+				break
+			}
 			case 1: {
 				if (content.charCodeAt(0) + content.charCodeAt(1) === 90) {
 					// collect custom properties
@@ -21,22 +31,12 @@
 					var name = content.substring(0, index)
 					var value = content.substring(index+1).trim()
 
-					return (properties[name] = value, '')
+					return (store[name] = value, '')
 				} else if (content.indexOf('var(')) {
 					// replace custom properties
 					return content.replace(/var\((.*)\)/g, replace)
 				}
 
-				break
-			}
-			case -1: {
-				// create store
-				properties = {}
-				break
-			}
-			case -2: {
-				// destroy store
-				properties = null
 				break
 			}
 		}